+++ Node.cpp
@@ -12,7 +12,7 @@ Node::Node(int elemIndex, int localNodeIndex) {
 Node::Node() {}
 
 
-long double Node::psi(long double x, long double y) {
+double Node::psi(double x, double y) {
     int sign[4][2] = {{-1, 1}, {1, 1}, {1, -1}, {-1, -1}};
 
     int signs = sign[this->localNodeIndex][0];
@@ -20,15 +20,15 @@ long double Node::psi(long double x, long double y) {
 
     pair<int, int> gridIndex = make_pair(this->elemIndex / Nx, this->elemIndex % Nx);  // exchange index
 
-    long double x0 = (gridIndex.second + 0.5) * h1;
-    long double y0 = (Ny * h2) - (gridIndex.first + 0.5) * h2;
-    long double s = 2 * (x - x0) / h1;
-    long double t = 2 * (y - y0) / h2;
+    double x0 = (gridIndex.second + 0.5) * h1;
+    double y0 = (Ny * h2) - (gridIndex.first + 0.5) * h2;
+    double s = 2 * (x - x0) / h1;
+    double t = 2 * (y - y0) / h2;
 
     return 0.25 * (1 + signs * s) * (1 + signt * t);
 }
 
-VectorXd Node::delpsi(long double x, long double y) {
+VectorXd Node::delpsi(double x, double y) {
     int sign[4][4] = {{-1, -1, 1, -1}, 
                       {1, 1, 1, 1},
                       {1, -1, -1, -1},
@@ -38,10 +38,10 @@ VectorXd Node::delpsi(long double x, long double y) {
 
     pair<int, int> gridIndex = make_pair(this->elemIndex / Nx, this->elemIndex % Nx);  // exchange index
 
-    long double x0 = (gridIndex.second + 0.5) * h1;
-    long double y0 = (Ny * h2) - (gridIndex.first + 0.5) * h2;
-    long double s = 2 * (x - x0) / h1;
-    long double t = 2 * (y - y0) / h2;
+    double x0 = (gridIndex.second + 0.5) * h1;
+    double y0 = (Ny * h2) - (gridIndex.first + 0.5) * h2;
+    double s = 2 * (x - x0) / h1;
+    double t = 2 * (y - y0) / h2;
 
     int sign1 = sign[this->localNodeIndex][0];
     int sign2 = sign[this->localNodeIndex][1];
